Turn of thread support
We don't share data between threads, so I hope we can get by with
disabling thread support. libssh with threads enabled rely on threading
being enables in mbedtls, but we have it disabled. In mbedtls, it is
fine to use mbedtls without thread support in a threaded environment as
long as there is no concurrent use of shared things between threads. We
never share anything between threads since the Actor run time ensures
only a single actor owns any crypto thing and we only run an actor by a
single worker thread at a time... this is explicitly stated in the
mbedtls docs but it doesn't read as clearly with libssh, so let's see.